refactor(embed): move the query-instruct seam out of gateway.ts — module-size ratchet + shared qwen3 family match
gateway.ts sits at its module-size ceiling (scripts/check-module-size.sh),
so the block reviewed in place moves verbatim into
src/core/ai/query-instruct.ts; gateway.ts grows by the import line only
(module-size-limits.tsv 4428→4429, noted). hybrid.ts' `qi=` knobs-ctx line
is the other +1 (2500→2501, noted).

isQwen3EmbeddingModel now lives in dims.ts (exported) and is shared by the
Matryoshka `dimensions` branch and the template. #4123 widened dims' match
to the cased hub form (`Qwen/Qwen3-Embedding-8B`), so a private copy here
would have silently broken this PR's own invariant ("same family match as
dims"). Consequence: hub-form ids get the template exactly where they get
Matryoshka dims. The lookalike negative test moves to an id outside the
family; one positive assertion pins the lockstep.

effectiveQueryInstruct(modelString?) — the cache-key feed with the
global-model default — lives in embedding.ts next to the other
gateway-wrapping accessors; query-instruct.ts stays gateway-free
(queryInstructFor(modelString), pure, never throws).

KEY_FILES.md: the GBRAIN_QUERY_INSTRUCT paragraph moves to a
query-instruct.ts entry; dims.ts entry names the shared predicate.
