+++ src/core/search/hybrid.ts
@@ -2050,10 +2050,7 @@ export async function hybridSearchCached(
     // bare hybridSearch does (opts.detail ?? autoDetectDetail(query)) so an
     // auto-detected `high` query keys like an explicit `high` one.
     detail: opts?.detail ?? autoDetectDetail(query),
-    // v=18 — fold the effective query-side instruction (the gateway's
-    // qwen3-embedding Instruct template) so rows written under different
-    // instruction settings never cross-serve.
-    queryInstruct: effectiveQueryInstruct(resolvedColCached.embeddingModel),
+    queryInstruct: effectiveQueryInstruct(resolvedColCached.embeddingModel), // v=20 qi= (see KnobsHashContext)
   });
 
   // Cache decision: opts.useCache (explicit) wins over global config; global
